Refactor context into it's own file
