+++ src/lib.rs
@@ -8,67 +8,16 @@ use sdl2::keyboard::Keycode;
 
 pub mod gl;
 pub mod error;
-pub mod builder;
+pub mod context;
 
-use error::AppResult;
 use gl::raw::types::*;
-
-struct Context {
-    sdl: sdl2::Sdl,
-    window: sdl2::video::Window,
-    event_pump: sdl2::EventPump,
-    gl_context: gl::GlContext,
-}
-
-impl Context {
-    // Initializes Window, Events, and Graphics Contexts
-    pub fn new() -> AppResult<Context> {
-        let sdl = sdl2::init()?;
-        let video = sdl.video()?;
-
-        let gl_attr = video.gl_attr();
-        gl_attr.set_context_profile(sdl2::video::GLProfile::Core);
-        gl_attr.set_context_version(3, 3);
-
-        debug_assert_eq!(gl_attr.context_version(), (3, 3));
-        // TODO: In house builder should just return this guy
-        let window = video.window("Window", 800, 600).opengl().build()?;
-
-        let gl_context = gl::GlContext::new(window.gl_create_context()?);
-        gl::raw::load_with(|name| video.gl_get_proc_address(name) as *const _);
-
-        let event_pump = sdl.event_pump()?;
-
-
-        debug_assert_eq!(gl_attr.context_version(), (3, 3));
-        Ok(Context {
-            sdl,
-            window,
-            event_pump,
-            gl_context,
-        })
-    }
-
-    pub fn window(&self) -> &sdl2::video::Window {
-        &self.window
-    }
-
-    pub fn present(&mut self) {
-        self.window.gl_swap_window();
-    }
-
-    pub fn poll_iter(&mut self) -> sdl2::event::EventPollIterator {
-        self.event_pump.poll_iter()
-    }
-
-    pub fn gl(&mut self) -> &gl::GlContext {
-        &self.gl_context
-    }
-}
+use context::ContextBuilder;
 
 pub fn run() -> error::AppResult<()> {
-    let mut ctx = Context::new()?;
-    debug_assert_eq!(ctx.window.subsystem().gl_attr().context_version(), (3, 3));
+    let mut ctx = ContextBuilder::default().build()?;
+
+    // TODO: Extract this into a Test
+    debug_assert_eq!(ctx.window().subsystem().gl_attr().context_version(), (3, 3));
     let vs = compile_shader(VS_SRC, gl::raw::VERTEX_SHADER);
     let fs = compile_shader(FS_SRC, gl::raw::FRAGMENT_SHADER);
